Handle multiple type parameter lists in value class methods
VCInlineMethods needed to be extended to this case.

Fixes #23266

Also rename previous i23266.scala to i23299.scala to reflect the proper issue for it.
