+++ compiler/test/dotc/pos-test-pickling.excludelist
@@ -139,7 +139,7 @@ parsercombinators-new-syntax.scala
 hylolib-deferred-given
 hylolib-cb
 hylolib
-i23266.scala
+i23299.scala
 
 # typecheckErrors method unpickling
 i21415.scala
