+++ tests/pos/i23266.scala
@@ -1,17 +1,8 @@
+//> using scala 3.7.0
 
-def kek(t: Table, ids: t.Id*) = ???
+class Foo(v: Any) extends AnyVal:
+  def bar[X](bar: X)[Y]: Any = v
 
-trait Table {
-  type Id = String
-}
-
-object Table1 extends Table {
-  val id: Id = "table1_id"
-}
-
-class Table2() extends Table {
-  val id: Id = "table2_id"
-}
-
-val x = kek(Table1, Table1.id)
-val y = kek(Table2(), Table2().id)
+@main def run: Unit =
+  val f = new Foo("lol")
+  println(f.bar[String]("")[Boolean])
